reorganize and refactor
