--- packages/desktop/lib/electron.ts
@@ -1,3 +1,0 @@
-import { IPlatform } from '@core/app'
-
-export const Electron: IPlatform = window['__ELECTRON__']
