+++ packages/desktop/lib/electron/apis/index.ts
@@ -1,0 +1,3 @@
+export { default as ElectronApi } from './electron.api'
+export { default as LedgerApi } from './ledger.api'
+export { default as WalletApi } from './wallet.api'
