+++ packages/desktop/lib/electron/index.ts
@@ -1,0 +1,5 @@
+import { IPlatform } from '@core/app/interfaces'
+
+export * from './utils'
+
+export const Electron: IPlatform = window['__ELECTRON__']
