+++ packages/desktop/lib/electron/managers/deep-link.manager.ts
