+++ packages/desktop/lib/electron/managers/keychain.manager.ts
