+++ packages/desktop/lib/electron/managers/nft-download.manager.ts
@@ -1,4 +1,4 @@
-import { getOrInitWindow } from '../../main'
+import { getOrInitWindow } from '../main'
 import { app, ipcMain, IpcMainInvokeEvent } from 'electron'
 import { download } from 'electron-dl'
