+++ packages/desktop/lib/electron/managers/notification.manager.ts
