+++ packages/desktop/lib/electron/managers/pincode.manager.ts
