+++ packages/desktop/lib/electron/menus/edit.menu.ts
