+++ packages/desktop/lib/electron/menus/help.menu.ts
