+++ packages/desktop/lib/electron/menus/index.ts
@@ -1,0 +1,6 @@
+export * from './context.menu'
+export * from './edit.menu'
+export * from './help.menu'
+export * from './menu'
+export * from './menu-state.constant'
+export * from './wallet.menu'
