+++ packages/desktop/lib/electron/menus/menu-state.constant.ts
