+++ packages/desktop/lib/electron/processes/index.ts
@@ -1,0 +1,1 @@
+export * from './ledger.process'
