+++ packages/desktop/lib/electron/utils/diagnostics.utils.ts
