+++ packages/desktop/lib/electron/utils/error.utils.ts
