+++ packages/desktop/lib/electron/utils/index.ts
@@ -1,0 +1,4 @@
+export * from './diagnostics.utils'
+export * from './error.utils'
+export * from './ledger.utils'
+export * from './os.utils'
