+++ packages/desktop/lib/electron/utils/os.utils.ts
